build: resolve 1 instance of -Wsign-compare
gcc-7 warns:

rfbl.cpp:1738:30: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if (((buf[12] & 0xF0) >> 4) != level || level != 2)

+++ lib/rfbl.cpp
@@ -1735,7 +1735,7 @@ int haproxy_intervene(int fd, unsigned int level, struct sockaddr_storage *ss)
 		return -1;
 	if (memcmp(buf, sig, sizeof(sig)) != 0)
 		return -1;
-	if (((buf[12] & 0xF0) >> 4) != level || level != 2)
+	if (static_cast<unsigned int>((buf[12] & 0xF0) >> 4) != level || level != 2)
 		return -1;
 	if ((buf[12] & 0xF) == 0)
 		return 0;
